src_1/Backend_Django/requirements.txt
asgiref==3.10.0
Django==5.2.7
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
